Add /W4 to get MS C++ to print extra warning that is reported as issue #8 - Use Py_hash_t for hash return type - long was losing info - Use Py_ssize_t for length and indices as python does - fix up tests and example to compile clean
git-svn-id: http://svn.code.sf.net/p/cxx/code@353 cb2f5747-f50d-0410-b093-b27f9c363ba5

+++ trunk/CXX/Demo/Python2/range.cxx
@@ -55,9 +55,9 @@ int range::sequence_length()
     return length();
 }
 
-Py::Object range::sequence_item(Py_ssize_t i) 
+Py::Object range::sequence_item( Py_ssize_t i ) 
 {
-    return Py::Int(item(i));
+    return Py::Int( item( static_cast<int>( i ) ) );
 }
 
 Py::Object range::sequence_concat( const Py::Object &j )
@@ -66,11 +66,10 @@ Py::Object range::sequence_concat( const Py::Object &j )
     return Py::asObject(extend(int(k)));
 }
 
-Py::Object range::sequence_slice(Py_ssize_t i, Py_ssize_t j)
+Py::Object range::sequence_slice( Py_ssize_t i, Py_ssize_t j )
 {
-    return Py::asObject(slice(i,j));
+    return Py::asObject( slice( static_cast<int>( i ), static_cast<int>( j ) ) );
 }
-
 
 Py::Object range::getattr( const char *name )
 {
